relocate convenience loaders to yt.loaders
