+++ setup.cfg
@@ -14,7 +14,7 @@ max-line-length=88
 exclude = doc,
           benchmarks,
           */api.py,                          # avoid spurious "unused import"
-          yt/loaders.py,                     # avoid spurious "unused import"
+          yt/convenience.py,                 # avoid spurious "unused import"
           */__init__.py,                     # avoid spurious "unused import"
           */__config__.py,                   # autogenerated
           yt/extern,                         # vendored libraries
